Improve thread locality of Header and Body data
Assuming decent thread scheduling and allocation; move header and body data to same NUMA node
